Introduce a CLI
- Add a new module to package a command line application.
- Move the core implementation to its own module.

The choice to make 2 modules comes from the fact that it is easier to package
the command line application this way. With a single module, it was not possible
to use maven-shade-plugin or appassembler-maven-plugin to add a dependency on a SLF4J
implementation. These plugins can only generate a classpath from the project dependency.
And I don't want to declare a logging dependency for the core implementation (Have not test
with 'optional' dependencies).

 send-notification [notifier] -m message -t title -i icon.png [-s subtitle]